% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{add_sitemap}
\alias{add_sitemap}
\title{Submit a sitemap.}
\usage{
add_sitemap(siteURL, feedpath)
}
\arguments{
\item{siteURL}{The URL of the website to delete. Must include protocol (http://).}

\item{feedpath}{The URL of the sitemap to submit. Must include protocol (http://).}
}
\value{
TRUE if successful, raises an error if not.
}
\description{
See here for details: https://developers.google.com/webmaster-tools/v3/sitemaps/submit
}
\seealso{
Other sitemap admin functions: 
\code{\link{delete_sitemap}()},
\code{\link{list_sitemaps}()}
}
\concept{sitemap admin functions}
