% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyAuth.R
\name{authReturnCode}
\alias{authReturnCode}
\title{Returns the authentication parameter "code" in redirected URLs}
\usage{
authReturnCode(session, securityCode)
}
\arguments{
\item{session}{A shiny session object}

\item{securityCode}{A random string to check the auth comes form the same origin.}
}
\value{
The Google auth token in the code URL parameter.
}
\description{
Checks the URL of the Shiny app to get the state and code URL parameters.
}
\keyword{internal}

