% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{checkGoogleAPIError}
\alias{checkGoogleAPIError}
\title{Get Google API errors}
\usage{
checkGoogleAPIError(req)
}
\arguments{
\item{a}{httr request}
}
\description{
Get Google API errors
}
\keyword{internal}

