% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{checkTokenAPI}
\alias{checkTokenAPI}
\title{Check API data token}
\usage{
checkTokenAPI(session, verbose = TRUE)
}
\arguments{
\item{shiny}{session.}
}
\value{
boolean if it works.
}
\description{
Check API data token
}
\seealso{
Other data fetching functions: \code{\link{delete_website}};
  \code{\link{list_sitemaps}}
}
\keyword{internal}

