% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{crawl_errors}
\alias{crawl_errors}
\title{Fetch a time-series of Googlebot crawl errors.}
\usage{
crawl_errors(siteURL, category = "all", platform = "all",
  latestCountsOnly = FALSE)
}
\arguments{
\item{siteURL}{The URL of the website to delete. Must include protocol (http://).}

\item{category}{Crawl error category. Defaults to 'all'}

\item{platform}{The user agent type. 'all', 'mobile', 'smartphoneOnly' or 'web'.}

\item{latestCountsOnly}{Default FALSE. Only the latest crawl error counts returned if TRUE.}
}
\value{
Dataframe of errors
}
\description{
Get a list of errors detected by Googlebot over time.
See here for details: https://developers.google.com/webmaster-tools/v3/urlcrawlerrorscounts/query
}
\details{
The timestamp is converted to a date as they are only available daily.

Category is one of: authPermissions, manyToOneRedirect, notFollowed, notFound,
  other, roboted, serverError, soft404.

  Platform is one of: mobile, smartphoneOnly or web.
}

