% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{delete_website}
\alias{delete_website}
\title{Deletes website in Search Console}
\usage{
delete_website(siteURL)
}
\arguments{
\item{siteURL}{The URL of the website to delete.}
}
\value{
TRUE if successful, raises an error if not.
}
\description{
Deletes website in Search Console
}

