% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{doHttrRequest}
\alias{doHttrRequest}
\title{Get URL content based on if its Shiny or local}
\usage{
doHttrRequest(url, request_type = "GET", the_body = NULL, params = NULL)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{request_type}{the type of httr request function: GET, POST, PUT, DELETE etc.}

\item{the_body}{body of POST request}

\item{params}{A named character vector of other parameters to add to request.}
}
\description{
This changes the auth type depending on if its local or on Shiny
}
\details{
Example of params: c(param1="foo", param2="bar")
}
\keyword{internal}

