% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyAuth.R
\name{get_google_token_shiny}
\alias{get_google_token_shiny}
\title{Does the Shiny authentication flow}
\usage{
get_google_token_shiny(session,
  sec_code = getOption("searchConsoleR.securitycode"))
}
\arguments{
\item{session}{A Shiny session object passed from shinyServer().}

\item{sec_code}{A random security code to check response is valid.}
}
\value{
A Google OAuth2 token
}
\description{
Does the Shiny authentication flow
}

