% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyAuth.R
\name{getShinyURL}
\alias{getShinyURL}
\title{Get the Shiny Apps URL.}
\usage{
getShinyURL(session)
}
\arguments{
\item{session}{The shiny session object.}
}
\value{
The URL of the Shiny App its called from.
}
\description{
Needed to for the redirect URL in Google Auth
}
\keyword{internal}

