% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection.R
\name{inspection}
\alias{inspection}
\title{Perform an inspection of URL index coverage}
\usage{
inspection(inspectionUrl, siteUrl, languageCode = NULL)
}
\arguments{
\item{inspectionUrl}{URL to inspect}

\item{siteUrl}{The URL of the property as defind in Search Console}

\item{languageCode}{Optional.  An IETF BCP-47 language code representing the requested language for translated issue messages e.g. "en-US" (the default)}
}
\value{
An R list containing the crawl information
}
\description{
Perform an inspection of URL index coverage
}
\examples{

\dontrun{

# siteUrl parameter has to be one of these
list_websites()

# get a URL from your website
inspection("https://code.markedmondson.me/searchConsoleR/",
           "sc-domain:code.markedmondson.me")
}
}
\seealso{
\url{https://developers.google.com/webmaster-tools/v1/urlInspection.index/inspect}
}
