% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is.valid.category.platform}
\alias{is.valid.category.platform}
\title{Checks valid parameters for Google Search Console}
\usage{
is.valid.category.platform(category, platform, include.all = FALSE)
}
\value{
TRUE if valid, raises an error if not.
}
\description{
The error type listings have certain categories/platform allowed.
}
\keyword{internal}

