% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{list_sitemaps}
\alias{list_sitemaps}
\title{Gets sitemap information for the URL supplied.}
\usage{
list_sitemaps(siteURL, session = NULL)
}
\arguments{
\item{siteURL}{The URL of the website to get sitemap information from. Must include protocol (http://).}

\item{session}{If in a Shiny session, supply its session object.}
}
\value{
A list of two dataframes: $sitemap with general info and $contents with sitemap info.
}
\description{
See here for details: https://developers.google.com/webmaster-tools/v3/sitemaps
}
\seealso{
Other data fetching functions: \code{\link{checkTokenAPI}};
  \code{\link{delete_website}}

Other sitemap admin functions: \code{\link{add_sitemap}};
  \code{\link{delete_sitemap}}
}

