% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{list_websites}
\alias{list_websites}
\title{Retrieves dataframe of websites user has in Search Console}
\usage{
list_websites()
}
\value{
a dataframe of siteUrl and permissionLevel
}
\description{
Retrieves dataframe of websites user has in Search Console
}

