% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reqParse.R
\name{parse_crawlerror_sample}
\alias{parse_crawlerror_sample}
\title{Parsing function for \code{\link{list_crawl_error_samples}}}
\usage{
parse_crawlerror_sample(x)
}
\arguments{
\item{x}{req$content from API response}
}
\description{
Parsing function for \code{\link{list_crawl_error_samples}}
}
\seealso{
Other parsing functions: \code{\link{parse_crawlerrors}};
  \code{\link{parse_errorsample_url}};
  \code{\link{parse_search_analytics}};
  \code{\link{parse_sitemaps}}
}
\keyword{internal}

