% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqParse.R
\name{parse_crawlerrors}
\alias{parse_crawlerrors}
\title{Parsing function for \code{\link{crawl_errors}}}
\usage{
parse_crawlerrors(x)
}
\arguments{
\item{x}{req$content from API response}
}
\description{
Parsing function for \code{\link{crawl_errors}}
}
\seealso{
Other parsing functions: \code{\link{parse_crawlerror_sample}},
  \code{\link{parse_errorsample_url}},
  \code{\link{parse_search_analytics}},
  \code{\link{parse_sitemaps}}
}
\keyword{internal}

