% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqParse.R
\name{parse_errorsample_url}
\alias{parse_errorsample_url}
\title{Parsing function for \code{\link{error_sample_url}}}
\usage{
parse_errorsample_url(x)
}
\arguments{
\item{x}{req$content from API response}
}
\description{
Parsing function for \code{\link{error_sample_url}}
}
\seealso{
Other parsing functions: \code{\link{parse_crawlerror_sample}},
  \code{\link{parse_crawlerrors}},
  \code{\link{parse_search_analytics}},
  \code{\link{parse_sitemaps}}
}
\concept{parsing functions}
\keyword{internal}
