% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reqParse.R
\name{parse_search_analytics}
\alias{parse_search_analytics}
\title{Parsing function for \code{\link{search_analytics}}}
\usage{
parse_search_analytics(x, dim, pn = TRUE)
}
\arguments{
\item{x}{req$content from API response}

\item{dim}{the dimensions passed from search_analytics}

\item{pn}{PrettyNames passed from search_analytics}
}
\description{
Parsing function for \code{\link{search_analytics}}
}
\seealso{
Other parsing functions: \code{\link{parse_crawlerror_sample}};
  \code{\link{parse_crawlerrors}};
  \code{\link{parse_errorsample_url}};
  \code{\link{parse_sitemaps}}
}
\keyword{internal}

