% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqParse.R
\name{parse_sitemaps}
\alias{parse_sitemaps}
\title{Parsing function for \code{\link{list_sitemaps}}}
\usage{
parse_sitemaps(x)
}
\arguments{
\item{x}{req$content from API response}
}
\description{
Parsing function for \code{\link{list_sitemaps}}
}
\seealso{
Other parsing functions: 
\code{\link{parse_crawlerror_sample}()},
\code{\link{parse_crawlerrors}()},
\code{\link{parse_errorsample_url}()},
\code{\link{parse_search_analytics}()}
}
\concept{parsing functions}
\keyword{internal}
