% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{searchconsole_DELETE}
\alias{searchconsole_DELETE}
\title{Create DELETE request}
\usage{
searchconsole_DELETE(url, params = NULL)
}
\arguments{
\item{url}{the url of the page to retrieve}
}
\description{
Make DELETE request to Search Console API.
}
\keyword{internal}

