% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{searchconsole_GET}
\alias{searchconsole_GET}
\title{Create GET request}
\usage{
searchconsole_GET(url, shiny_access_token = NULL, to_json = TRUE,
  params = NULL)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{to_json}{whether to convert response contents to JSON or
leave as character string}

\item{params}{A named character vector of other parameters to add to request.}

\item{use_auth}{logical; indicates if authorization should be used, defaults
to \code{FALSE} if \code{url} implies public visibility and \code{TRUE}
otherwise}
}
\description{
Make GET request to Search Console API.
}
\keyword{internal}

