% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/http_requests.R
\name{searchconsole_PUT}
\alias{searchconsole_PUT}
\title{Create PUT request}
\usage{
searchconsole_PUT(url, the_body, params = NULL)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{the_body}{body of PUT request}
}
\description{
Make PUT request to Search Console API.
}
\keyword{internal}

