% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyAuth.R
\name{shinygaGetToken}
\alias{shinygaGetToken}
\title{Returns the authentication Token.}
\usage{
shinygaGetToken(code, redirect.uri,
  client.id = getOption("searchConsoleR.webapp.client_id"),
  client.secret = getOption("searchConsoleR.webapp.client_secret"))
}
\arguments{
\item{code}{The code returned from a successful Google authentication.}

\item{redirect.uri}{Where a user will go after authentication.}

\item{client.id}{From the Google API console.}

\item{client.secret}{From the Google API console.}
}
\value{
A list including the token needed for Google API requests.
}
\description{
Once a user browses to ShinyGetTokenURL and is redirected back with request
shinygaGetToken takes that code and returns a token needed for Google APIs
Uses the same client.id and client.secret as ShinyGetTokenURL.
}
\keyword{internal}

