% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyAuth.R
\name{shinygaGetTokenURL}
\alias{shinygaGetTokenURL}
\title{Returns the Google authentication URL}
\usage{
shinygaGetTokenURL(state, redirect.uri,
  client.id = getOption("SearchConsoleR.webapp.client_id"),
  client.secret = getOption("SearchConsoleR.webapp.client_secret"),
  scope = getOption("SearchConsoleR.scope"))
}
\arguments{
\item{state}{A random string used to check auth is from same origin.}

\item{redirect.uri}{Where a user will go after authentication.}

\item{client.id}{From the Google API console.}

\item{client.secret}{From the Google API console.}

\item{scope}{What Google API service to get authentication for.}
}
\value{
The URL for authentication.
}
\description{
The URL a user authenticates the Shiny app on.
}
\keyword{internal}

