% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scr_auth.R
\docType{data}
\name{Authentication}
\alias{Authentication}
\title{R6 environment to store credentials}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb10cec55c8> 
 - attr(*, "name")= chr "Authentication_generator"
}}
\usage{
Authentication
}
\description{
Stores local authentication details
}
\keyword{datasets}

