% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scr_auth.R
\docType{data}
\name{Authentication}
\alias{Authentication}
\title{In search of an object ot refer from within package and shiny session}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7f8118ab88d8> 
 - attr(*, "name")= chr "Authentication_generator"
}}
\usage{
Authentication
}
\description{
In search of an object ot refer from within package and shiny session
}
\keyword{datasets}

