% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{delete_sitemap}
\alias{delete_sitemap}
\title{Delete a sitemap.}
\usage{
delete_sitemap(siteURL, feedpath)
}
\arguments{
\item{siteURL}{The URL of the website you are deleting the sitemap from. Must include protocol (http://).}

\item{feedpath}{The URL of the sitemap to delete. Must include protocol (http://).}
}
\value{
TRUE if successful, raises an error if not.
}
\description{
See here for details: https://developers.google.com/webmaster-tools/v3/sitemaps/delete
}

