% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{list_websites}
\alias{list_websites}
\title{Retrieves dataframe of websites user has in Search Console}
\usage{
list_websites()
}
\value{
a dataframe of siteUrl and permissionLevel
}
\description{
Retrieves dataframe of websites user has in Search Console
}
\seealso{
Other search console website functions: \code{\link{add_website}},
  \code{\link{delete_website}}
}

