% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyAuth.R
\name{shinygaGetTokenURL}
\alias{shinygaGetTokenURL}
\title{Returns the Google authentication URL}
\usage{
shinygaGetTokenURL(redirect.uri = getShinyURL(session),
  state = getOption("searchConsoleR.securitycode"),
  client.id = getOption("searchConsoleR.webapp.client_id"),
  client.secret = getOption("searchConsoleR.webapp.client_secret"),
  scope = getOption("searchConsoleR.scope"))
}
\arguments{
\item{redirect.uri}{App URL, from getShinyURL(session)}

\item{state}{A random string used to check auth is from same origin.}

\item{client.id}{From the Google API console.}

\item{client.secret}{From the Google API console.}

\item{scope}{What Google API service to get authentication for.}
}
\value{
The URL for authentication.
}
\description{
The URL a user authenticates the Shiny app on.
}
\seealso{
Other Shiny.Authentication: \code{\link{getShinyURL}}
}

